RecommendationService/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
python-jose
